requirements.txt
Flask==2.0.2
gevent==21.12.0
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
redis==4.1.2
Flask-JWT-Extended==4.3.1